backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "claude-agent-sdk>=0.1.29",
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "python-dotenv>=1.2.1",
]
